% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_function.R
\name{calculate_and_compare_volume}
\alias{calculate_and_compare_volume}
\title{Calculate and Compare Computed and Theoretical Volume of a Champagne Glass

This function calculates the volume of a champagne glass using numerical integration
and compares it with a theoretical value.}
\usage{
calculate_and_compare_volume(
  f_vectorized_for_integration,
  lower,
  upper,
  true_formula
)
}
\arguments{
\item{f_vectorized_for_integration}{A vectorized function representing the profile of the champagne glass.}

\item{lower}{The lower bound for the integral.}

\item{upper}{The upper bound for the integral.}

\item{true_formula}{A character string representing the formula for the theoretical volume.}
}
\value{
A list containing the computed volume (V_computed) and the theoretical volume (V_true).
}
\description{
Calculate and Compare Computed and Theoretical Volume of a Champagne Glass

This function calculates the volume of a champagne glass using numerical integration
and compares it with a theoretical value.
}
\examples{
f_vectorized_for_integration <- function(x) { f_for_loop(x) ** 2 }
lower_bound <- 10
upper_bound <- 20
true_formula <- "8 * pi * (80 + 5 * log2(2) + 88 * log2(6) + 4 * (11 * log(6) - 5) * log2(2) - 96 * log2(6)) * log2(2)"
result <- calculate_and_compare_volume(f_vectorized_for_integration, lower_bound, upper_bound, true_formula)
print(result)
}
\author{
Nicolas Zwickert
}
