% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_for_loop.R
\name{f_for_loop}
\alias{f_for_loop}
\title{Vectorized Champagne Glass Profile Function using For Loop}
\usage{
f_for_loop(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the result of applying `f_non_vectorized` to each element of `x`.
}
\description{
Vectorized Champagne Glass Profile Function using For Loop
}
\author{
Nicolas Zwickert
}
