% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_purrr_f.R
\name{f_purrr}
\alias{f_purrr}
\title{Vectorized Champagne Glass Profile Function using purrr

This function vectorizes `f_non_vectorized` using the `map_dbl` function from the `purrr` package.}
\usage{
f_purrr(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the result of applying `f_non_vectorized` to each element of `x`.
}
\description{
Vectorized Champagne Glass Profile Function using purrr

This function vectorizes `f_non_vectorized` using the `map_dbl` function from the `purrr` package.
}
\examples{
f_purrr(c(1, 5, 10))
}
\author{
Nicolas Zwickert
}
