% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sapply_f.R
\name{f_sapply}
\alias{f_sapply}
\title{Vectorized Champagne Glass Profile Function using sapply}
\usage{
f_sapply(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the result of applying `f_non_vectorized` to each element of `x`.
}
\description{
Vectorized Champagne Glass Profile Function using sapply
}
\examples{
f_sapply(10)
}
\author{
Nicolas Zwickert
}
