% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_vectorized_f.R
\name{f_vectorized}
\alias{f_vectorized}
\title{Directly Vectorized Champagne Glass Profile Function

This function directly vectorizes `f_non_vectorized` using `Vectorize`.
It allows `f_non_vectorized` to accept vector inputs directly.}
\usage{
f_vectorized(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the result of applying `f_non_vectorized` to each element of `x`.
}
\description{
Directly Vectorized Champagne Glass Profile Function

This function directly vectorizes `f_non_vectorized` using `Vectorize`.
It allows `f_non_vectorized` to accept vector inputs directly.
}
\examples{
f_vectorized(c(1, 5, 10))
}
\author{
Nicolas Zwickert
}
