% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_for_integration.R
\name{f_vectorized_for_integration}
\alias{f_vectorized_for_integration}
\title{Vectorized Integration for Champagne Glass Profile}
\usage{
f_vectorized_for_integration(x)
}
\arguments{
\item{x}{A numeric vector representing x-coordinates for which the champagne glass profile is calculated.}
}
\value{
A numeric vector with the champagne glass profile values for each x-coordinate, suitable for integration.
}
\description{
Vectorized Integration for Champagne Glass Profile
}
\examples{
# Assuming f_for_loop is defined
x_values <- seq(0, 20, by = 0.1)
integrated_values <- f_vectorized_for_integration(x_values)
plot(x_values, integrated_values, type = 'l')
}
\author{
Nicolas Zwickert
}
