% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_volume_function.R
\name{optimize_volume_function}
\alias{optimize_volume_function}
\title{Volume Optimization for a Given Range}
\usage{
optimize_volume_function(b, target_volume)
}
\arguments{
\item{b}{The upper limit for the integration, which determines the level of Champagne in the glass.}

\item{target_volume}{A numeric value specifying the target volume for calculations.}
}
\value{
The absolute difference between the computed volume of the Champagne glass and the target volume.
}
\description{
Volume Optimization for a Given Range
}
\examples{
optimize_volume_function(15,200)
}
\author{
Nicolas Zwickert
}
