% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_party.R
\name{simulate_party}
\alias{simulate_party}
\title{Champagne Party Simulation}
\usage{
simulate_party()
}
\value{
A numeric vector with two elements: the first being the total liters of Champagne served, and the second the total number of bottles needed.
}
\description{
Champagne Party Simulation
}
\examples{
# Example assuming f_for_loop and necessary functions are defined
party_simulation <- simulate_party()
print(party_simulation)
}
\author{
Nicolas Zwickert
}
