% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_function.R
\name{V_function}
\alias{V_function}
\title{Square of the Champagne Glass Profile Function}
\usage{
V_function(x)
}
\arguments{
\item{x}{A numeric vector representing the x-coordinates at which to calculate the square of the champagne glass profile function.}
}
\value{
A numeric vector representing the square of the champagne glass profile values at each x-coordinate.
}
\description{
Square of the Champagne Glass Profile Function
}
\examples{
x <- seq(0, 20, by = 0.1)
}
\author{
Nicolas Zwickert
}
