% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_function_uniroot.R
\name{volume_function_uniroot}
\alias{volume_function_uniroot}
\title{Volume Calculation for uniroot}
\usage{
volume_function_uniroot(b, target_volume)
}
\arguments{
\item{b}{Upper limit for the integration, which defines the level of Champagne in the glass.}

\item{target_volume}{A numeric value specifying the target volume for calculations.}
}
\value{
The difference between the calculated volume of the Champagne glass and the target volume.
}
\description{
Volume Calculation for uniroot
}
\examples{
x <- seq(0, 20, by = 0.1)
V_function(x)
volume_function_uniroot(15,200)
}
\author{
Nicolas Zwickert
}
