% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_direct_vectorized.R
\name{f_direct_vectorized}
\alias{f_direct_vectorized}
\title{Alternative Vectorized Champagne Glass Profile Function}
\usage{
f_direct_vectorized(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the result of applying `f_non_vectorized` to each element of `x`.
}
\description{
Alternative Vectorized Champagne Glass Profile Function
}
\examples{
f_direct_vectorized(10)
}
\author{
Nicolas Zwickert
}
