% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_for_loop.R
\name{f_for_loop}
\alias{f_for_loop}
\title{Vectorized Champagne Glass Profile - For Loop}
\usage{
f_for_loop(x)
}
\arguments{
\item{x}{Numeric vector of x-coordinates.}
}
\value{
Numeric vector of glass heights at the given x-coordinates.
}
\description{
Vectorized Champagne Glass Profile - For Loop
}
\examples{
f_for_loop(c(1, 5, 10))
}
\author{
Nicolas Zwickert
}
