% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_non_vectorized.R
\name{f_non_vectorized}
\alias{f_non_vectorized}
\title{Non-Vectorized Champagne Glass Profile Function

This function represents the profile of a Champagne glass as a piecewise function.
It returns specific values based on the input `x`.}
\usage{
f_non_vectorized(x)
}
\arguments{
\item{x}{A numeric value representing the position on the x-axis.}
}
\value{
A numeric value representing the height of the glass at position `x`.
}
\description{
Non-Vectorized Champagne Glass Profile Function

This function represents the profile of a Champagne glass as a piecewise function.
It returns specific values based on the input `x`.
}
\author{
Nicolas Zwickert
}
