% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_non_vectorized.R
\name{f_non_vectorized}
\alias{f_non_vectorized}
\title{Non-Vectorized Champagne Glass Profile}
\usage{
f_non_vectorized(x)
}
\arguments{
\item{x}{Numeric value, the x-coordinate.}
}
\value{
Numeric value, the height of the glass at the given x-coordinate.
}
\description{
Non-Vectorized Champagne Glass Profile
}
\examples{
f_non_vectorized(16)
}
\author{
Nicolas Zwickert
}
