% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sapply_f.R
\name{f_sapply}
\alias{f_sapply}
\title{Vectorized Champagne Glass Profile Function using sapply

This function vectorizes `f_non_vectorized` using `sapply`. It is used to apply `f_non_vectorized` to each element of a numeric vector.}
\usage{
f_sapply(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the result of applying `f_non_vectorized` to each element of `x`.
}
\description{
Vectorized Champagne Glass Profile Function using sapply

This function vectorizes `f_non_vectorized` using `sapply`. It is used to apply `f_non_vectorized` to each element of a numeric vector.
}
\examples{
f_sapply(c(1, 5, 10))
}
\author{
Nicolas Zwickert
}
