% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_vectorized_f.R
\name{f_vectorized}
\alias{f_vectorized}
\title{Directly Vectorized Champagne Glass Profile Function}
\usage{
f_vectorized(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with the result of applying `f_non_vectorized` to each element of `x`.
}
\description{
Directly Vectorized Champagne Glass Profile Function
}
\examples{
f_vectorized(10)
}
\author{
Nicolas Zwickert
}
