% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_vectorized_for_integration.R
\name{f_vectorized_for_integration}
\alias{f_vectorized_for_integration}
\title{Vectorized Integration for Champagne Glass Profile}
\usage{
f_vectorized_for_integration(x)
}
\arguments{
\item{x}{A numeric vector representing x-coordinates for which the champagne glass profile is calculated.}
}
\value{
A numeric vector with the champagne glass profile values for each x-coordinate, suitable for integration.
}
\description{
Vectorized Integration for Champagne Glass Profile
}
\examples{
x <- seq(0, 20, by = 0.1)
integrated_values <- f_vectorized_for_integration(x)
}
\author{
Nicolas Zwickert
}
