% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_volume_function.R
\name{optimize_volume_function}
\alias{optimize_volume_function}
\title{Volume Optimization for a Given Range}
\usage{
optimize_volume_function(b)
}
\arguments{
\item{b}{The upper limit for the integration, which determines the level of Champagne in the glass.}
}
\value{
The absolute difference between the computed volume of the Champagne glass and the target volume.
}
\description{
Volume Optimization for a Given Range
}
\examples{
V_function <- function(x) { f_vectorized_for_integration(x) ** 2 }
target_volume <- 200  # Target volume in cubic centimeters
b_optimize <- optimize(optimize_volume_function, c(10, 25))$minimum
print(b_optimize)
}
\author{
Nicolas Zwickert
}
