% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_function_uniroot.R
\name{volume_function_uniroot}
\alias{volume_function_uniroot}
\title{Volume Calculation for uniroot}
\usage{
volume_function_uniroot(b)
}
\arguments{
\item{b}{Upper limit for the integration, which defines the level of Champagne in the glass.}
}
\value{
The difference between the calculated volume of the Champagne glass and the target volume.
}
\description{
Volume Calculation for uniroot
}
\examples{
V_function <- function(x) { f_vectorized_for_integration(x) ** 2 }
target_volume <- 200  # Target volume in cubic centimeters
b_uniroot <- uniroot(volume_function_uniroot, c(10, 25))$root
print(b_uniroot)
}
\author{
Nicolas Zwickert
}
